- More color improvements - Add texture utilities for loading and resizing - Streaming texture improvements - Fix stencil dying when using too much text - Virtual list grid fixes - Add steamworks storage

+++ rawrbox.math/include/rawrbox/math/utils/color.hpp
@@ -6,6 +6,6 @@
 namespace rawrbox {
 	class ColorUtils {
 	public:
-		static std::vector<uint8_t> setChannels(size_t channels, size_t target, size_t width, size_t height, const std::vector<uint8_t>& buffer);
+		static std::vector<uint8_t> setChannels(uint8_t channels, uint8_t target, size_t width, size_t height, const std::vector<uint8_t>& buffer);
 	};
 } // namespace rawrbox
