+++ rawrbox.render/include/rawrbox/render/textures/image.hpp
@@ -18,10 +18,10 @@ namespace rawrbox {
 		explicit TextureImage(const std::filesystem::path& filePath, const std::vector<uint8_t>& buffer, bool useFallback = true);
 		explicit TextureImage(const uint8_t* buffer, int bufferSize, bool useFallback = true); // TO BE DECODED
 
-		explicit TextureImage(const rawrbox::Vector2u& size, const uint8_t* buffer, int channels = 4);
-		explicit TextureImage(const rawrbox::Vector2u& size, const std::vector<uint8_t>& buffer, int channels = 4);
+		explicit TextureImage(const rawrbox::Vector2u& size, const uint8_t* buffer, uint8_t channels = 4);
+		explicit TextureImage(const rawrbox::Vector2u& size, const std::vector<uint8_t>& buffer, uint8_t channels = 4);
 
-		explicit TextureImage(const rawrbox::Vector2u& size, int channels = 4);
+		explicit TextureImage(const rawrbox::Vector2u& size, uint8_t channels = 4);
 
 		TextureImage(const TextureImage&) = delete;
 		TextureImage(TextureImage&&) = delete;
