+++ rawrbox.steamworks/include/rawrbox/steamworks/workshop/manager.hpp
@@ -12,6 +12,7 @@
 #include <string>
 
 namespace rawrbox {
+
 	enum class WorkshopStatus : uint32_t {
 		NONE = 0U,
 		INSTALLED,
@@ -103,8 +104,8 @@ namespace rawrbox {
 		// -------------
 
 		// QUERIES ---
-		static void queryUserMods(const std::function<void(std::vector<SteamUGCDetails_t>)>& callback, EUserUGCList type = k_EUserUGCList_Subscribed, const std::vector<std::string>& tags = {}, size_t page = 1);
-		static void queryMods(const std::function<void(std::vector<SteamUGCDetails_t>)>& callback, EUGCQuery type = k_EUGCQuery_RankedByVote, const std::vector<std::string>& tags = {}, size_t page = 1);
+		static void queryUserMods(const std::function<void(std::vector<SteamUGCDetails_t>)>& callback, EUserUGCList type = k_EUserUGCList_Subscribed, const std::vector<std::string>& tags = {}, uint32_t page = 1U);
+		static void queryMods(const std::function<void(std::vector<SteamUGCDetails_t>)>& callback, EUGCQuery type = k_EUGCQuery_RankedByVote, const std::vector<std::string>& tags = {}, uint32_t page = 1U);
 		//-------------
 
 		// UPLOAD ------
